New module _models.py
