+++ src/vessqc/__init__.py
@@ -2,11 +2,11 @@ __version__ = "0.8.0"
 
 from ._widget import (
     ExampleQWidget,
-    Segment,
     _save_npy,
     _load_npy,
     _build_filename,
 )
+from ._models import Segment
 
 __all__ = (
     "ExampleQWidget",
